refactor: Break out update logic to separate module
